+++ kolibri/plugins/coach/assets/src/composables/useResourceSelection.js
